fix: Delay CSS generation till after hugo_stats.json is availabe
If CSS generation happens before the hugo_stats.json is built,
important CSS class will be missing from the generation.  This
can be avoided by using `resources.PostProcess` in the pipeline
generating the CSS output.

Also switch to using RelPermalink, as otherwise the url will be
incorrect with the change.